requirements.txt
streamlit>=1.33
pandas>=2.0
python-docx>=1.1
pdfplumber>=0.11
pypdf>=4.0
reportlab>=4.0
